requirements.txt
# Pinned runtime dependencies for TRACE
openai==2.16.0
anthropic==0.79.0
google-genai==1.62.0
jsonschema==4.26.0
traitlets==5.14.3
